main file
Smart Face Recognition System

The primary goal of this project is to enable real-time detection and recognition of individuals using a face recognition system. When a person is detected, the system checks if their information is stored in the database. If it is, their name will appear on the screen above their face. By clicking on the highlighted rectangle around the person’s face and pressing Enter, the system will display detailed information about them, similar to an ID card. Additionally, an alert voice notification will sound whenever an unknown person is detected.